sadfasfdasdasdaf

+++ libft/libft.h
@@ -74,7 +74,7 @@ char				*ft_itoa(int n);
 void				ft_putchar(char c);
 void				ft_putstr(char const *str);
 void				ft_putendl(char const *str);
-void				ft_putnbr(int nb);
+void				ft_putnbr(long int nb);
 void				ft_putchar_fd(char c, int fd);
 void				ft_putstr_fd(char const *str, int fd);
 void				ft_putendl_fd(char const *str, int fd);
